Extract class AllAtOnceKeywordsFinder to dedicated files.
