+++ src/task.hpp
@@ -16,6 +16,7 @@
 #include <tuple>
 #include <type_traits>
 
+#include "allatoncekeywordsfinder.h"
 #include "recurrence_instance_data.hpp"
 #include "task_date_time.hpp"
 #include "taskproperty.hpp"
@@ -159,36 +160,6 @@ class BatchTasksManager {
     [[nodiscard]]
     bool execVerb(const QString &verb, const TaskWarriorExecutor &executor,
                   const QString &after_ids = {}) const;
-};
-
-/// @brief Keywords set by user. `task` finds all IDs which contain ALL
-/// keywords.
-class AllAtOnceKeywordsFinder {
-  public:
-    explicit AllAtOnceKeywordsFinder(QStringList keywords);
-
-    [[nodiscard]]
-    bool readIds(const TaskWarriorExecutor &executor);
-
-    /// @returns std::nullopt if it was no filter applied, empty string if it
-    /// was no task found by keywords, or task ids list.
-    [[nodiscard]]
-    const auto &getIds() const
-    {
-        return m_ids;
-    }
-
-    /// @returns true if nothing was found, i.e. it should be displayed empty
-    /// results.
-    [[nodiscard]]
-    bool isNotFound() const
-    {
-        return getIds().has_value() && getIds()->isEmpty();
-    }
-
-  private:
-    QStringList m_user_keywords;
-    std::optional<QString> m_ids;
 };
 
 /// @brief Commands to read tasks list. Tasks will have particulary filled data
